Add capacitors directly across pH probes.
This fixes a problem with 60 Hz (or 50 Hz) pickup, when the wires are
long or the noise is coupling to the water. Also replace the 1k resistors
to the shield with zero-ohm resistors, to further reduce that noise.

This updates the board to rev C. Also check in PDF schematics, BOM,
and Gerbers.